More code cleanup and move function to test helper file
